+++ frontend/packages/volto-light-theme/src/types.d.ts
@@ -9,26 +9,46 @@ type headerAction = {
   '@id': string;
   title: string;
   href: Array<hrefType>;
+  openInNewTab: boolean;
 };
 
-type footerLink = {
+export type Link = {
   '@id': string;
   title: string;
   href: Array<hrefType>;
+  openInNewTab: boolean;
 };
 
-type footerLogo = {
+export type Logo = {
   '@id': string;
   title: string;
   logo: Image;
   href: Array<hrefType>;
+  openInNewTab: boolean;
+};
+
+export type footerLogo = {
+  '@id': string;
+  title: string;
+  logo: Image;
+  href: Array<hrefType>;
+  openInNewTab: boolean;
+};
+
+export type iconLink = {
+  '@id': string;
+  title: string;
+  icon: Image;
+  href: Array<hrefType>;
+  openInNewTab: boolean;
 };
 
 export type SiteHeaderSettings = {
-  logo: string;
-  complementary_logo: string;
+  logo: Image & { data?: string };
+  complementary_logo: Image & { data?: string };
   intranet_flag: string;
   header_actions: Array<headerAction>;
+  has_intranet_header: boolean;
 };
 
 export type SiteThemeSettings = {
@@ -41,10 +61,22 @@ export type SiteThemeSettings = {
 };
 
 export type SiteFooterSettings = {
-  footer_links: Array<footerLink>;
   footer_logos: Array<footerLogo>;
   footer_logos_container_width: string;
   footer_logos_size: string;
+  footer_address: string;
+  footer_column_left_header: string;
+  footer_column_left: Array<Link>;
+  footer_column_middle_header: string;
+  footer_column_middle: Array<Link>;
+  footer_column_right_header: string;
+  footer_column_right: Array<Link>;
+  footer_logo: Image & { data?: string };
+  footer_links: Array<Link>;
+};
+
+export type PloneGobrSocialMediaSettings = {
+  social_links: Array<iconLink>;
 };
 
 export type CustomInheritBehavior<T> = {
@@ -58,5 +90,6 @@ export type CustomInheritBehavior<T> = {
 declare module '@plone/types' {
   export interface Content {
     footer_logos: Array<footerLogo>;
+    footer_links: Array<Link>;
   }
 }